df: print "total" in grand total's target field if no source field
If no source field is used, then df would print "-" into the
target column. Instead, df should print "total":

  df --output=pcent,target --total '.'
  Use% Mounted on
   75% /media/sdb5
   75% total

* src/df.c (main): Pass "total" as the mount point to get_dev if
the SOURCE_FIELD is not among the columns to output.
* tests/df/df-output.sh: Change the test to ensure the content of
the target field of the grand total line: if the source field is
present, then the target should be "-", else the target field should
be "target".
* NEWS (Changes in behavior): Enhance the exiting NEWS entry.
* doc/coreutils.texi (df invocation): Document the content of the
source and target field in the grand total line.