df: ensure that the mount list contains a valid fstype with --output
The function read_file_system_list only ensures a valid file
system type if the calling parameter is true. With the new
option --output, there is an additional condition needed:
when the FSTYPE_FIELD is used.

* src/df.c (main): Add another condition to the need_fs_type
parameter of read_file_system_list whether the FSTYPE_FIELD is
used or not.