numfmt: fix negative number support
* src/numfmt.c (simple_round_from_zero): A new function equivalent
to ceil() for positive numbers, but symmetrically rounds away from
zero for negative numbers.  Set this as the default rounding method.
(absld): A new function to get the absolute value of a long double,
without the need for libm.  Use this throughout when comparing
the magnitude of floating point values.
* tests/misc/numfmt.pl: Add new tests for negative number formats,
and rounding issues.