cp: -Z: install: fix setting of system default context
install(1) by default sets the context for target files
to their system default.  So disable the older method
to do this when -Z specified, to avoid redundant processing.

Also allow the -Z option (that doesn't now take a context argument)
to select between the new and old context restoration behavior.

Document the differences and details for how context restoration
is done in new and old methods, with a view disabling the
old method entirely in future.

* src/install.c (main): When -Z specified, disable the old
setdefaultcon() method.
Also ensure it's disabled for --preserve-context.