cp: -Z: refactor setting of security context
Refactor to two separate functions that set the
process and file security context.  These are then
used to consistently fail and/or warn as required.

* src/copy.c (set_process_security_ctx): A new function,
refactored to set the default context from the source file,
or with the type adjusted as per the system default for
the destination path.
(set_file_security_ctx): A new function refactored to
set the security context of an existing file, either based on
the process context or the default system context for a path.
(copy_internal): Use the refactored functions to simplify
error handling and consistently fail or warn as needed.
(copy_reg): Likewise.